\name{mywaffler}
\alias{mywaffler}
\title{Create Waffle Charts}
\description{
  This function generates waffle charts from a data frame. Each waffle chart represents the proportion of categories
  based on a given numeric value.
}
\usage{
mywaffler(df, category_col, value_col, color_col = NULL)
}
\arguments{
  \item{df}{A data frame containing the data. It must have at least two columns: one for categories and one for values.}
  \item{category_col}{A string representing the name of the column in \code{df} that contains the category names.}
  \item{value_col}{A string representing the name of the column in \code{df} that contains the numeric values corresponding to each category.}
  \item{color_col}{An optional string representing the name of the column in \code{df} that defines the colors of the chart's segments. If \code{NULL}, colors are automatically generated.}
}
\value{
  This function produces a series of waffle charts. No return value.
}
\author{
  Your Name
}
\keyword{visualization}
